+++ CHANGELOG.md
@@ -22,6 +22,8 @@ The format is based on [Keep a Changelog].
 
   See [#27].
 
+* The function `selectrum-read-directory-name` is no longer available.
+
 ### Features
 * You can now give a prefix argument to
   `selectrum-insert-current-candidate` to insert the candidate at a
@@ -95,8 +97,8 @@ The format is based on [Keep a Changelog].
   insert file names into the minibuffer (using
   `file-name-at-point-functions`) and you are able to use shortcuts
   like `//` or `~/` ([#50], [#52]).
-* When reading directories using `read-directory-name` the default is
-  sorted to the top instead of inserting it ([#50]).
+* When reading directories and the default is already in the prompt,
+  it gets selected so you can immediately submit it ([#126], [#127]).
 * In `read-file-name`, when a default is provided (for example in the
   `dired-do-rename` command), we actually use it as the initial
   contents of the minibuffer, which allows you to have convenient
@@ -218,6 +220,8 @@ The format is based on [Keep a Changelog].
 [#113]: https://github.com/raxod502/selectrum/issues/113
 [#118]: https://github.com/raxod502/selectrum/pull/118
 [#125]: https://github.com/raxod502/selectrum/pull/125
+[#126]: https://github.com/raxod502/selectrum/issues/126
+[#127]: https://github.com/raxod502/selectrum/pull/127
 [raxod502/ctrlf#41]: https://github.com/raxod502/ctrlf/issues/41
 
 ## 1.0 (released 2020-03-23)